requirements.txt
langchain
langchain-groq
langchain-community
python-dotenv
pandas